Revert "METRON-2175 Introduce HBase Connection Abstractions for HBase 2.0.2 (nickwallen) closes apache/metron#1456"
This reverts commit ef3a76c799d48856d8e4999226ec9392f298a2f7.
