+++ metron-platform/metron-hbase/metron-hbase-common/src/main/java/org/apache/metron/hbase/HBaseProjectionCriteria.java
@@ -40,16 +40,12 @@ public class HBaseProjectionCriteria implements Serializable {
 
   public static class ColumnMetaData implements Serializable {
 
-    private byte[] columnFamily;
+    private byte[]  columnFamily;
     private byte[] qualifier;
 
     public ColumnMetaData(String columnFamily, String qualifier) {
-      this(columnFamily.getBytes(), qualifier.getBytes());
-    }
-
-    public ColumnMetaData(byte[] columnFamily, byte[] qualifier) {
-      this.columnFamily = columnFamily;
-      this.qualifier = qualifier;
+      this.columnFamily = columnFamily.getBytes();
+      this.qualifier = qualifier.getBytes();
     }
 
     public byte[] getColumnFamily() {
