+++ metron-platform/metron-hbase/metron-hbase-common/src/main/java/org/apache/metron/hbase/client/HBaseClient.java
@@ -17,114 +17,321 @@
  *  limitations under the License.
  *
  */
+
 package org.apache.metron.hbase.client;
 
-import org.apache.hadoop.hbase.client.Durability;
-import org.apache.hadoop.hbase.client.Result;
-import org.apache.metron.hbase.ColumnList;
-import org.apache.metron.hbase.HBaseProjectionCriteria;
+import static org.apache.commons.collections4.CollectionUtils.size;
 
 import java.io.Closeable;
 import java.io.IOException;
+import java.lang.invoke.MethodHandles;
+import java.util.ArrayList;
 import java.util.List;
+import org.apache.hadoop.conf.Configuration;
+import org.apache.hadoop.hbase.client.Durability;
+import org.apache.hadoop.hbase.client.Get;
+import org.apache.hadoop.hbase.client.HTableInterface;
+import org.apache.hadoop.hbase.client.Increment;
+import org.apache.hadoop.hbase.client.Mutation;
+import org.apache.hadoop.hbase.client.Put;
+import org.apache.hadoop.hbase.client.Result;
+import org.apache.hadoop.hbase.client.ResultScanner;
+import org.apache.hadoop.hbase.client.Scan;
+import org.apache.hadoop.hbase.util.Bytes;
+import org.apache.metron.hbase.TableProvider;
+import org.apache.metron.hbase.ColumnList;
+import org.apache.metron.hbase.HBaseProjectionCriteria;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
  * A client that interacts with HBase.
  */
-public interface HBaseClient extends Closeable {
+public class HBaseClient implements Closeable {
+
+  private static final Logger LOG = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
 
   /**
-   * Enqueues a 'get' request that will be submitted when {@link #getAll()} is called.
-   * @param rowKey The row key to be retrieved.
+   * The batch of queued Mutations.
    */
-  void addGet(byte[] rowKey, HBaseProjectionCriteria criteria);
+  List<Mutation> mutations;
 
   /**
-   * Submits all pending get operations and returns the result of each.
-   * @return The result of each pending get request.
+   * The batch of queued Gets.
    */
-  Result[] getAll();
+  List<Get> gets;
 
   /**
-   * Clears all pending get operations.
+   * The HBase table this client interacts with.
    */
-  void clearGets();
+  private HTableInterface table;
+
+  public HBaseClient(TableProvider provider, final Configuration configuration, final String tableName) {
+    this.mutations = new ArrayList<>();
+    this.gets = new ArrayList<>();
+    try {
+      this.table = provider.getTable(configuration, tableName);
+    } catch (Exception e) {
+      String msg = String.format("Unable to open connection to HBase for table '%s'", tableName);
+      LOG.error(msg, e);
+      throw new RuntimeException(msg, e);
+    }
+  }
 
   /**
-   * Scans an entire table returning all row keys as a List of Strings.
-   *
-   * <p><b>**WARNING**:</b> Do not use this method unless you're absolutely crystal clear about the performance
-   * impact. Doing full table scans in HBase can adversely impact performance.
-   *
-   * @return List of all row keys as Strings for this table.
-   */
-  List<String> scanRowKeys() throws IOException;
-
-  /**
-   * Scans the table and returns each result.
-   *
-   * <p><b>**WARNING**:</b> Do not use this method unless you're absolutely crystal clear about the performance
-   * impact. Doing full table scans in HBase can adversely impact performance.
-   *
-   * @return The results from the scan.
-   * @throws IOException
-   */
-  Result[] scan(int numRows) throws IOException;
-
-  /**
-   * Enqueues a {@link org.apache.hadoop.hbase.client.Mutation} such as a put or
-   * increment.  The operation is enqueued for later execution.
-   *
-   * @param rowKey     The row key of the Mutation.
-   * @param cols       The columns affected by the Mutation.
-   */
-  void addMutation(byte[] rowKey, ColumnList cols);
-
-  /**
-   * Enqueues a {@link org.apache.hadoop.hbase.client.Mutation} such as a put or
-   * increment.  The operation is enqueued for later execution.
+   * Add a Mutation such as a Put or Increment to the batch.  The Mutation is only queued for
+   * later execution.
    *
    * @param rowKey     The row key of the Mutation.
    * @param cols       The columns affected by the Mutation.
    * @param durability The durability of the mutation.
    */
-  void addMutation(byte[] rowKey, ColumnList cols, Durability durability);
+  public void addMutation(byte[] rowKey, ColumnList cols, Durability durability) {
+
+    if (cols.hasColumns()) {
+      Put put = createPut(rowKey, cols, durability);
+      mutations.add(put);
+    }
+
+    if (cols.hasCounters()) {
+      Increment inc = createIncrement(rowKey, cols, durability);
+      mutations.add(inc);
+    }
+
+    if (mutations.isEmpty()) {
+      mutations.add(new Put(rowKey));
+    }
+  }
 
   /**
-   * Enqueues a {@link org.apache.hadoop.hbase.client.Mutation} such as a put or
-   * increment.  The operation is enqueued for later execution.
+   * Adds a Mutation such as a Put or Increment with a time to live.  The Mutation is only queued
+   * for later execution.
    *
    * @param rowKey           The row key of the Mutation.
    * @param cols             The columns affected by the Mutation.
    * @param durability       The durability of the mutation.
    * @param timeToLiveMillis The time to live in milliseconds.
    */
-  void addMutation(byte[] rowKey, ColumnList cols, Durability durability, Long timeToLiveMillis);
+  public void addMutation(byte[] rowKey, ColumnList cols, Durability durability, Long timeToLiveMillis) {
+
+    if (cols.hasColumns()) {
+      Put put = createPut(rowKey, cols, durability, timeToLiveMillis);
+      mutations.add(put);
+    }
+
+    if (cols.hasCounters()) {
+      Increment inc = createIncrement(rowKey, cols, durability, timeToLiveMillis);
+      mutations.add(inc);
+    }
+
+    if (mutations.isEmpty()) {
+      Put put = new Put(rowKey);
+      put.setTTL(timeToLiveMillis);
+      mutations.add(put);
+    }
+  }
 
   /**
-   * Ensures that all pending mutations have completed.
+   * Remove all queued Mutations from the batch.
+   */
+  public void clearMutations() {
+    mutations.clear();
+  }
+
+  /**
+   * Submits all queued Mutations.
+   * @return The number of mutation submitted.
+   */
+  public int mutate() {
+    int mutationCount = mutations.size();
+    Object[] result = new Object[mutationCount];
+    try {
+      table.batch(mutations, result);
+      mutations.clear();
+
+    } catch (Exception e) {
+      String msg = String.format("'%d' HBase write(s) failed on table '%s'", size(mutations), tableName(table));
+      LOG.error(msg, e);
+      throw new RuntimeException(msg, e);
+    }
+
+    return mutationCount;
+  }
+
+  /**
+   * Adds a Get to the batch.
    *
-   * @return The number of operations completed.
+   * @param rowKey   The row key of the Get
+   * @param criteria Defines the columns/families that will be retrieved.
    */
-  int mutate();
+  public void addGet(byte[] rowKey, HBaseProjectionCriteria criteria) {
+    Get get = new Get(rowKey);
+
+    if (criteria != null) {
+      criteria.getColumnFamilies().forEach(cf -> get.addFamily(cf));
+      criteria.getColumns().forEach(col -> get.addColumn(col.getColumnFamily(), col.getQualifier()));
+    }
+
+    // queue the get
+    this.gets.add(get);
+  }
 
   /**
-   * Clears all pending mutations.
+   * Clears all queued Gets from the batch.
    */
-  void clearMutations();
+  public void clearGets() {
+    gets.clear();
+  }
 
   /**
-   * Delete a record by row key.
+   * Submit all queued Gets.
    *
-   * @param rowKey The row key to delete.
+   * @return The Result of each queued Get.
    */
-  void delete(byte[] rowKey);
+  public Result[] getAll() {
+    try {
+      Result[] results = table.get(gets);
+      gets.clear();
+      return results;
+
+    } catch (Exception e) {
+      String msg = String.format("'%d' HBase read(s) failed on table '%s'", size(gets), tableName(table));
+      LOG.error(msg, e);
+      throw new RuntimeException(msg, e);
+    }
+  }
 
   /**
-   * Delete a column or set of columns by row key.
-   *
-   * @param rowKey The row key to delete.
-   * @param columnList The set of columns to delete.
+   * Close the table.
    */
-  void delete(byte[] rowKey, ColumnList columnList);
+  @Override
+  public void close() throws IOException {
+    if(table != null) {
+      table.close();
+    }
+  }
+
+  /**
+   * Creates an HBase Put.
+   *
+   * @param rowKey     The row key.
+   * @param cols       The columns to put.
+   * @param durability The durability of the put.
+   */
+  private Put createPut(byte[] rowKey, ColumnList cols, Durability durability) {
+    Put put = new Put(rowKey);
+    put.setDurability(durability);
+    addColumns(cols, put);
+    return put;
+  }
+
+  /**
+   * Creates an HBase Put.
+   *
+   * @param rowKey           The row key.
+   * @param cols             The columns to put.
+   * @param durability       The durability of the put.
+   * @param timeToLiveMillis The TTL in milliseconds.
+   */
+  private Put createPut(byte[] rowKey, ColumnList cols, Durability durability, long timeToLiveMillis) {
+    Put put = new Put(rowKey);
+    put.setDurability(durability);
+    put.setTTL(timeToLiveMillis);
+    addColumns(cols, put);
+    return put;
+  }
+
+  /**
+   * Adds the columns to the Put
+   *
+   * @param cols The columns to add.
+   * @param put  The Put.
+   */
+  private void addColumns(ColumnList cols, Put put) {
+    for (ColumnList.Column col : cols.getColumns()) {
+
+      if (col.getTs() > 0) {
+        put.add(col.getFamily(), col.getQualifier(), col.getTs(), col.getValue());
+
+      } else {
+        put.add(col.getFamily(), col.getQualifier(), col.getValue());
+      }
+    }
+  }
+
+  /**
+   * Creates an HBase Increment for a counter.
+   *
+   * @param rowKey     The row key.
+   * @param cols       The columns to include.
+   * @param durability The durability of the increment.
+   */
+  private Increment createIncrement(byte[] rowKey, ColumnList cols, Durability durability) {
+    Increment inc = new Increment(rowKey);
+    inc.setDurability(durability);
+    cols.getCounters().forEach(cnt -> inc.addColumn(cnt.getFamily(), cnt.getQualifier(), cnt.getIncrement()));
+    return inc;
+  }
+
+  /**
+   * Creates an HBase Increment for a counter.
+   *
+   * @param rowKey     The row key.
+   * @param cols       The columns to include.
+   * @param durability The durability of the increment.
+   */
+  private Increment createIncrement(byte[] rowKey, ColumnList cols, Durability durability, long timeToLiveMillis) {
+    Increment inc = new Increment(rowKey);
+    inc.setDurability(durability);
+    inc.setTTL(timeToLiveMillis);
+    cols.getCounters().forEach(cnt -> inc.addColumn(cnt.getFamily(), cnt.getQualifier(), cnt.getIncrement()));
+    return inc;
+  }
+
+  /**
+   * Returns the name of the HBase table.
+   * <p>Attempts to avoid any null pointers that might be encountered along the way.
+   * @param table The table to retrieve the name of.
+   * @return The name of the table
+   */
+  private static String tableName(HTableInterface table) {
+    String tableName = "null";
+    if(table != null) {
+      if(table.getName() != null) {
+        tableName = table.getName().getNameAsString();
+      }
+    }
+    return tableName;
+  }
+
+  /**
+   * Puts a record into the configured HBase table synchronously (not batched).
+   */
+  public void put(String rowKey, String columnFamily, String columnQualifier, String value)
+      throws IOException {
+    Put put = new Put(Bytes.toBytes(rowKey));
+    put.addColumn(Bytes.toBytes(columnFamily), Bytes.toBytes(columnQualifier),
+        Bytes.toBytes(value));
+    table.put(put);
+  }
+
+  /**
+   * Scans an entire table returning all row keys as a List of Strings.
+   *
+   * <p>
+   * <b>**WARNING**:</b> Do not use this method unless you're absolutely crystal clear about the performance
+   * impact. Doing full table scans in HBase can adversely impact performance.
+   *
+   * @return List of all row keys as Strings for this table.
+   */
+  public List<String> readRecords() throws IOException {
+    Scan scan = new Scan();
+    ResultScanner scanner = table.getScanner(scan);
+    List<String> rows = new ArrayList<>();
+    for (Result r = scanner.next(); r != null; r = scanner.next()) {
+      rows.add(Bytes.toString(r.getRow()));
+    }
+    return rows;
+  }
+
 }
