+++ pom.xml
@@ -91,6 +91,7 @@
         <base_flux_version>1.0.1</base_flux_version>
         <base_kafka_version>0.10.0</base_kafka_version>
         <base_hadoop_version>2.7.1</base_hadoop_version>
+        <base_hbase_version>1.1.1</base_hbase_version>
         <base_flume_version>1.5.2</base_flume_version>
         <!-- full dependency versions -->
         <global_accumulo_version>1.8.0</global_accumulo_version>
@@ -105,12 +106,14 @@
         <global_pcap_version>1.7.1</global_pcap_version>
         <global_kafka_version>0.10.0.1</global_kafka_version>
         <global_hadoop_version>${base_hadoop_version}</global_hadoop_version>
+        <global_hbase_version>${base_hbase_version}</global_hbase_version>
         <global_flume_version>${base_flume_version}</global_flume_version>
         <global_elasticsearch_version>5.6.14</global_elasticsearch_version>
         <global_json_simple_version>1.1.1</global_json_simple_version>
         <global_metrics_version>3.0.2</global_metrics_version>
         <global_junit_version>4.12</global_junit_version>
         <global_guava_version>17.0</global_guava_version>
+        <global_hbase_guava_version>12.0</global_hbase_guava_version>
         <global_json_schema_validator_version>2.2.5</global_json_schema_validator_version>
         <global_slf4j_version>1.7.7</global_slf4j_version>
         <global_opencsv_version>3.7</global_opencsv_version>
@@ -136,13 +139,12 @@
         <global_jacoco_version>0.8.3</global_jacoco_version>
         <argLine></argLine>
     </properties>
+
     <profiles>
         <profile>
             <id>HDP-3.1</id>
             <properties>
                 <hdp_version>3.1.0.0</hdp_version>
-                <global_hbase_version>2.0.2</global_hbase_version>
-                <global_hbase_guava_version>17.0</global_hbase_guava_version>
                 <global_zeppelin_version>0.8.0</global_zeppelin_version>
             </properties>
         </profile>
@@ -153,8 +155,6 @@
             </activation>
             <properties>
                 <hdp_version>2.6.5.0</hdp_version>
-                <global_hbase_version>1.1.1</global_hbase_version>
-                <global_hbase_guava_version>12.0</global_hbase_guava_version>
                 <global_zeppelin_version>0.8.0</global_zeppelin_version>
             </properties>
         </profile>
@@ -166,12 +166,11 @@
                 <global_storm_kafka_version>1.2.2</global_storm_kafka_version>
                 <global_storm_version>${base_storm_version}.${hdp_version}-${build_number}</global_storm_version>
                 <global_kafka_version>${base_kafka_version}.${hdp_version}-${build_number}</global_kafka_version>
-                <global_hbase_version>1.1.1</global_hbase_version>
-                <global_hbase_guava_version>12.0</global_hbase_guava_version>
                 <global_zeppelin_version>0.8.0</global_zeppelin_version>
             </properties>
         </profile>
     </profiles>
+
     <dependencyManagement>
         <dependencies>
             <dependency>
